♻️ Refactor: 예약 현황페이지 서버 컴포넌트로 전환
